menu toggle fixes
