+++ app/static/js/adapters.js
@@ -6,7 +6,6 @@ import {
     toggleMenu,
     stopAllffmpegs
 } from './modules/a_controlls.js';
-// import { showNewAdapterForm } from './modules/a_form.js';
 
 document.addEventListener('DOMContentLoaded', function () {
     updateAdapters();
@@ -26,157 +25,16 @@ document.addEventListener('DOMContentLoaded', function () {
         } else if (target.classList.contains('menu-button')) {
             const menuId = target.getAttribute('data-menu-id');
             toggleMenu(menuId);
+        } else if (target.classList.contains('toggle-program-details')) {
+            const programTitle = target.getAttribute('data-program-title');
+            toggleProgramDetails(programTitle);
+        } else if (target.classList.contains('toggle-url-list')) {
+            const urlListId = target.getAttribute('data-url-list-id');
+            toggleUrlList(urlListId);
         }
     });
 });
 
-function addEventListeners() {
-    document.getElementById('stop-all-ffmpegs').addEventListener('click', stopAllffmpegs);
-    // document.getElementById('new-adapter-form').addEventListener('click', showNewAdapterForm);
-
-
-    // Add event listeners for start buttons
-    document.querySelectorAll('.start-ffmpeg').forEach(button => {
-        button.addEventListener('click', function () {
-            const adapterId = this.getAttribute('data-adapter-id');
-            startFFmpeg(adapterId);
-        });
-    });
-
-    // Add event listeners for stop buttons
-    document.querySelectorAll('.stop-ffmpeg').forEach(button => {
-        button.addEventListener('click', function () {
-            const adapterId = this.getAttribute('data-adapter-id');
-            stopFFmpeg(adapterId);
-        });
-    });
-
-    // Add event listeners for scan buttons
-    document.querySelectorAll('.scan-adapter').forEach(button => {
-        button.addEventListener('click', function () {
-            const adapterId = this.getAttribute('data-adapter-id');
-            scanAdapter(adapterId);
-        });
-    });
-
-    // Add event listeners for delete buttons
-    document.querySelectorAll('.delete-adapter').forEach(button => {
-        button.addEventListener('click', function () {
-            const adapterId = this.getAttribute('data-adapter-id');
-            deleteAdapter(adapterId);
-        });
-    });
-
-    // Add event listeners for menu buttons
-    document.querySelectorAll('.menu-button').forEach(button => {
-        button.addEventListener('click', function () {
-            const menuId = this.getAttribute('data-menu-id');
-            toggleMenu(menuId);
-        });
-    });
-}
-
-
-function updateChannelSelection(programId) {
-    const channelCheckbox = document.getElementById(`channel-${programId}`);
-    const streamCheckboxes = document.querySelectorAll(`[data-channel="${programId}"]`);
-
-    // Set all stream checkboxes based on channel checkbox state
-    streamCheckboxes.forEach(cb => cb.checked = channelCheckbox.checked);
-}
-
-function updateStreamSelection(programId) {
-    const channelCheckbox = document.getElementById(`channel-${programId}`);
-    const streamCheckboxes = document.querySelectorAll(`[data-channel="${programId}"]`);
-
-    // Update the channel checkbox based on the stream checkboxes
-    const allStreamsSelected = Array.from(streamCheckboxes).every(cb => cb.checked);
-    const anyStreamSelected = Array.from(streamCheckboxes).some(cb => cb.checked);
-
-    if (allStreamsSelected) {
-        channelCheckbox.checked = true;
-    } else if (!anyStreamSelected) {
-        channelCheckbox.checked = false;
-    }
-}
-
-function toggleChannel(event, programId) {
-    event.preventDefault(); // Prevent default behavior of the label
-    const channelCheckbox = document.getElementById(`channel-${programId}`);
-    channelCheckbox.checked = !channelCheckbox.checked;
-    updateChannelSelection(programId);
-}
-
-function toggleStream(event, streamId) {
-    event.preventDefault(); // Prevent default behavior of the label
-    const streamCheckbox = document.getElementById(`video-${streamId}`) || document.getElementById(`audio-${streamId}`) || document.getElementById(`subtitle-${streamId}`);
-    streamCheckbox.checked = !streamCheckbox.checked;
-    const channelId = streamCheckbox.getAttribute('data-channel');
-    updateStreamSelection(channelId);
-}
-
-
-function saveSelection(adapterId) {
-    const selectedChannels = {};
-
-    // Collect selected streams
-    const streamCheckboxes = document.querySelectorAll('input[type="checkbox"][data-id]');
-    streamCheckboxes.forEach(cb => {
-        if (cb.checked) {
-            const streamId = cb.dataset.id;
-            if (!cb.dataset.channel)
-                return;
-            const channelId = cb.dataset.channel;
-            if (!selectedChannels[channelId]) {
-                selectedChannels[channelId] = {
-                    video: [],
-                    audio: [],
-                    subtitle: []
-                };
-            }
-            if (cb.id.startsWith('video-')) {
-                selectedChannels[channelId].video.push(streamId);
-            } else if (cb.id.startsWith('audio-')) {
-                selectedChannels[channelId].audio.push(streamId);
-            } else if (cb.id.startsWith('subtitle-')) {
-                selectedChannels[channelId].subtitle.push(streamId);
-            }
-        }
-    });
-
-    console.log(`SELECTED: \n${selectedChannels}`)
-    fetch(`/adapters/${adapterId}/save`, {
-            method: 'POST',
-            headers: {
-                'Content-Type': 'application/json'
-            },
-            body: JSON.stringify({
-                channels: selectedChannels
-            })
-        })
-        .then(response => response.json())
-        .then(data => {
-            showPopup(data.msg, data.status);
-            updateAdapters(adapterId)
-        })
-        .catch(error => showPopup(error, "error"));
-}
-
-function setAdapterControlDisplay(id, isDisplay) {
-    const adapterDiv = document.getElementById(`adapter-${id}`);
-    adapterDiv.querySelector(`#scan-button-${id}`).style.display = isDisplay ? 'inline' : 'none';
-    adapterDiv.querySelector(`#start-form-${id}`).style.display = isDisplay ? 'inline' : 'none';
-    adapterDiv.querySelector(`#stop-form-${id}`).style.display = isDisplay ? 'inline' : 'none';
-    adapterDiv.querySelector(`#delete-form-${id}`).style.display = isDisplay ? 'inline' : 'none';
-}
-
-function cancelSelection(adapterId) {
-    const scanSection = document.getElementById(`scan-section-${adapterId}`);
-    // Restore other controls
-    setAdapterControlDisplay(adapterId, true);
-    // Clear scan results
-    scanSection.innerHTML = '';
-}
 export function updateAdapters(adapterId = null) {
     const url = adapterId ? `/get_adapter/${adapterId}/` : '/get_adapters/';
 
@@ -193,7 +51,6 @@ export function updateAdapters(adapterId = null) {
                     updateAdapterDiv(id, adapter);
                 });
             }
-            //addEventListeners();
         })
         .catch(error => showPopup(error, "error"));
 }
@@ -232,7 +89,7 @@ function updateAdapterDiv(adapterId, adapter) {
     const selectedChannelsHtml = selectedPrograms.length ?
         selectedPrograms.map(program => `
             <li>
-                <a href="javascript:void(0);" onclick="toggleProgramDetails('${program.title}-${adapterId}')">
+                <a href="javascript:void(0);" class="toggle-program-details" data-program-title="${program.title}-${adapterId}">
                     ${program.title}
                 </a>
                 <div id="program-details-${program.title}-${adapterId}" class="program-details" style="display:none; margin-left: 20px;">
@@ -263,7 +120,7 @@ function updateAdapterDiv(adapterId, adapter) {
     // Adapter type and UDP URLs (with Expand/Collapse for SPTS)
     const udpUrlsHtml = adapter.type === 'SPTS' ? `
         <div>
-            <a href="javascript:void(0);" onclick="toggleUrlList('${adapterId}-urls')">UDP URLs: &#11206;</a>
+            <a href="javascript:void(0);" class="toggle-url-list" data-url-list-id="${adapterId}-urls">UDP URLs: &#11206;</a>
             <ul id="${adapterId}-urls" style="display:none;">
                 ${adapter.udp_urls.map(urlConfig => `
                     <li>
